Add user crud and refactor
